Update script to include custom fields
We use the "Custom Fields" powerup and I've added the parameter to the request to include custom fields. This adds a new "customFields" list to the master JSON that includes details about the fields you've created, as well as a "customFieldItems" list to each card that includes the values for any custom fields attached to those cards.